+++ go.work
@@ -1,7 +1,7 @@
 go 1.22.0
 
 use (
-	./scr-ops
+	./ops
 	./superchain
 	./validation
 )
