+++ justfile
@@ -4,41 +4,41 @@ alias l := lint-all
 
 # Adding a chain
 add-chain:
-	go run ./scr-ops add-new-chain
-	go run ./scr-ops check-rollup-config
-	go run ./scr-ops compress-genesis
-	go run ./scr-ops check-genesis
+	go run ./ops add-new-chain
+	go run ./ops check-rollup-config
+	go run ./ops compress-genesis
+	go run ./ops check-genesis
 	just codegen
 
 # Promote a standard candidate chain to a standard chain, example: just promote-to-standard 10
 promote-to-standard CHAIN:
-	go run ./scr-ops promote-to-standard --chain-id={{CHAIN}}
+	go run ./ops promote-to-standard --chain-id={{CHAIN}}
 	just codegen
 
 # Generate auto-generated files
-codegen: clean-scr-ops
+codegen: clean-ops
 	CODEGEN=true go run superchain/internal/codegen/main.go
 
 # Lint all Go code
 lint-all:
-	golangci-lint run superchain/... validation/... scr-ops/... --fix
+	golangci-lint run superchain/... validation/... ops/... --fix
 
 # Test all Go code
-test-all: test-scr-ops test-superchain test-validation
+test-all: test-ops test-superchain test-validation
 
-# Test all Go code in the scr-ops module
-test-scr-ops:
+# Test all Go code in the ops module
+test-ops:
 	# We separate the first test from the rest because it generates artifacts
 	# Which need to exist before the remaining tests run.
-	TEST_DIRECTORY=./scr-ops go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run TestScrOps_Main
-	TEST_DIRECTORY=./scr-ops/... go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run '[^TestScrOps_Main]'
+	TEST_DIRECTORY=./ops go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run TestScrOps_Main
+	TEST_DIRECTORY=./ops/... go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run '[^TestScrOps_Main]'
 
 # Test all Go code in the superchain module
-test-superchain: clean-scr-ops
+test-superchain: clean-ops
 	TEST_DIRECTORY=./superchain go run gotest.tools/gotestsum@latest --format testname
 
 # Unit test all Go code in the validation module, and do not run validation checks themselves
-test-validation: clean-scr-ops
+test-validation: clean-ops
 	TEST_DIRECTORY=./validation/... go run gotest.tools/gotestsum@latest --format testname -- -run '[^TestValidation|^TestPromotion|^TestGenesisAllocs]'
 
 # Runs validation checks for any chain whose config changed
@@ -65,18 +65,18 @@ promotion-test:
 validate-standard-candidate CHAIN_ID:
 	TEST_DIRECTORY=./validation go run gotest.tools/gotestsum@latest --format testname -- -run='Promotion/.+\({{CHAIN_ID}}\)$' -count=1
 
-# Clean test files generated by the scr-ops tooling
-clean-scr-ops:
+# Clean test files generated by the ops tooling
+clean-ops:
 	rm -f superchain/configs/sepolia/testchain_*.toml
 	rm -f superchain/extra/sepolia/testchain_*.json.gz
 	rm -rf -- validation/genesis/validation-inputs/*-test/
 
 # Tidying all go.mod files
-tidy-all: tidy-scr-ops tidy-superchain tidy-validation
+tidy-all: tidy-ops tidy-superchain tidy-validation
 
-# Tidy the scr-ops go.mod file
-tidy-scr-ops:
-	cd scr-ops && go mod tidy
+# Tidy the ops go.mod file
+tidy-ops:
+	cd ops && go mod tidy
 
 # Tidy the superchain go.mod file
 tidy-superchain:
