+++ ops/cmd/add-new-chain.go
@@ -13,9 +13,9 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/ethereum-optimism/optimism/op-e2e/bindings"
 	"github.com/ethereum-optimism/optimism/op-service/retry"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/config"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/flags"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/utils"
+	"github.com/ethereum-optimism/superchain-registry/ops/config"
+	"github.com/ethereum-optimism/superchain-registry/ops/flags"
+	"github.com/ethereum-optimism/superchain-registry/ops/utils"
 	"github.com/ethereum-optimism/superchain-registry/superchain"
 	"github.com/ethereum-optimism/superchain-registry/validation"
 	"github.com/ethereum-optimism/superchain-registry/validation/genesis"
