+++ ops/cmd/compress-genesis.go
@@ -12,8 +12,8 @@ import (
 	"runtime"
 
 	"github.com/ethereum-optimism/optimism/op-service/jsonutil"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/flags"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/utils"
+	"github.com/ethereum-optimism/superchain-registry/ops/flags"
+	"github.com/ethereum-optimism/superchain-registry/ops/utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core"
