+++ ops/cmd/promote-to-standard.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 	"runtime"
 
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/config"
-	"github.com/ethereum-optimism/superchain-registry/scr-ops/flags"
+	"github.com/ethereum-optimism/superchain-registry/ops/config"
+	"github.com/ethereum-optimism/superchain-registry/ops/flags"
 	"github.com/ethereum-optimism/superchain-registry/superchain"
 	"github.com/urfave/cli/v2"
 )
